#3 Classes movements and rename
