+++ SlnfUpdater/Processor/ProjectFile/ProjectFileProcessor.cs
@@ -36,7 +36,7 @@ namespace SlnfUpdater.Processor.ProjectFile
 
     //public sealed class CleanupRootsProcessor : ProjectFileProcessor
     //{
-    //    private readonly SlnfJsonStructured _structuredJson;
+    //    private readonly SlnfJsonStructured _slnf;
         
     //    /// <summary>
     //    /// projects from slnf which have to references from other slnf's projects.
@@ -44,15 +44,15 @@ namespace SlnfUpdater.Processor.ProjectFile
     //    private List<string> _existingRoots = [];
 
     //    public CleanupRootsProcessor(
-    //        SlnfJsonStructured structuredJson
+    //        SlnfJsonStructured slnf
     //        )
     //    {
-    //        if (structuredJson is null)
+    //        if (slnf is null)
     //        {
-    //            throw new ArgumentNullException(nameof(structuredJson));
+    //            throw new ArgumentNullException(nameof(slnf));
     //        }
 
-    //        _structuredJson = structuredJson;
+    //        _slnf = slnf;
     //    }
 
     //    public override void ProcessProjectFromSlnf(
@@ -71,15 +71,15 @@ namespace SlnfUpdater.Processor.ProjectFile
 
     public sealed class ScanForChangesProcessor : ProjectFileProcessor
     {
-        private readonly SlnfJsonStructured _structuredJson;
+        private readonly Slnf _slnf;
         private readonly SearchReferenceContext _context;
 
         public ScanForChangesProcessor(
-            SlnfJsonStructured structuredJson
+            Slnf slnf
             )
         {
-            _structuredJson = structuredJson ?? throw new ArgumentNullException(nameof(structuredJson));
-            _context = structuredJson.BuildSearchReferenceContext();
+            _slnf = slnf ?? throw new ArgumentNullException(nameof(slnf));
+            _context = slnf.BuildSearchReferenceContext();
         }
 
         public void ApplyAndSave()
@@ -89,8 +89,8 @@ namespace SlnfUpdater.Processor.ProjectFile
                 return;
             }
 
-            _context.ApplyChangesTo(_structuredJson);
-            _structuredJson.SerializeToItsFile();
+            _context.ApplyChangesTo(_slnf);
+            _slnf.SerializeToItsFile();
         }
 
         public string BuildResultMessage()
